### 2018-05-18 #218 @nbenn
Submission: infx
Data from large-scale biological experiments, managed by the [openBIS](https://sis.id.ethz.ch/software/openbis.html) framework, is made accessible directly from R. The main focus is on data from [InfectX](http://www.systemsx.ch/projects/research-technology-and-development-projects/completed-rtd-projects/infectx-2010-2013)/[TargetInfectX](http://www.systemsx.ch/projects/research-technology-and-development-projects/targetinfectx), which includes many image-based high throughput screening experiments, investigating the human infectome of several bacterial and viral pathogens.

    ```
    Package: infx
    Title: OpenBIS API Access to the InfectX Data Repository
    Version: 0.1.0
    Authors@R: person("Nicolas", "Bennett",
                      email = "nicolas.bennett@stat.math.ethz.ch",
                      role = c("aut", "cre"))
    Description: The Open Source Biology Information System (openBIS) is a
        general purpose framework for management, annotation and publication
        of large data sets that arise from biological experiments. By making
        the JSON-RPC based openBIS API available to R, image-based high
        throughput screening data as generated by the InfectX/TargetInfectX
        projects can be browsed, searched and downloaded directly from R.
        Currently, several kinome-wide RNA interference screens performed on
        HeLa cells in presence of a selection of bacterial and viral pathogens
        and using oligo libraries form multiple vendors are available. Further
        genome-wide screens are forthcoming. The full data obtained from these
        experiments is accessible, including raw microscopy images, object
        segmentation masks, single cell feature data generated by CellProfiler
        and infection scoring data, alongside rich meta data and quality
        control data.
    URL:
        https://nbenn.github.io/infx,
        http://www.infectx.ch,
        https://www.targetinfectx.ch,
        https://openbis.elnlims.ch
        R.matlab,
        assertthat,
        curl,
        progress,
        crayon,
        magick,
        base64enc
    License: GPL-3 | file LICENSE
        rvest,
        RColorBrewer
    RoxygenNote: 6.0.1.9000
    ```

    https://github.com/nbenn/infx

    Data retrieval: Data and associated meta data can be browsed, queried and downloaded.

    Anyone interested in developing methodology for analysis of image-based high throughput screening data. Unlike data sets from the [Broad Bioimage Benchmark Collection](https://data.broadinstitute.org/bbbc/image_sets.html), the full data stack, including raw data, meta data, quality control and analysis (e.g. feature extraction at single cell level) results are accessible. Furthermore, screens are available with replicates, in pooled as well as unpooled fashion and using libraries from several manufacturers, making this an extensive data resource.

    No.
