### 2018-05-25 #218 @sckott
thanks for your submission @nbenn

doing editor checks now, and haven't seen any documentation in your package on how to create an account? It seems one can create an API token with `login_openbis`, but how do you get a username and password?  Would be good to add this info to the package as well.
