### 2018-06-05 #218 @nbenn
@sckott, sorry for the delayed response on my side.

There is currently no way for a user to create an account him/herself. Accounts can only be created by the service admins. All publicly available data can be accessed using publicly released login credentials. To make this more straightforward, I added these public login credentials as default arguments to `login_openbis()` such that

```r
tok <- login_openbis()
```

now yields a valid login token associated with this public user account.
