### 2018-06-11 #218 @sckott
### Editor checks:

---

#### Editor comments

Thanks for your submission @nbenn   !!

- last i checked `\donttest` examples will still be run on CRAN, I think in the last few years they've started to run those examples - use `\dontrun` if you want them to not be run on CRAN
- the test suite takes quite a long time to run, i'd think about if there's any way to shorten them as i'm pretty sure CRAN maintainers will complain about the time, what I get:

```r
* checking tests ...
  Running ‘test-1-utils.R’ [7s/30s]
  Running ‘test-2-objects.R’ [37s/95s]
  Running ‘test-3-datasets.R’ [22s/130s]
  Running ‘test-4-features.R’ [12s/36s]
  Running ‘test-5-images.R’ [11s/169s]
  Running ‘test-6-files.R’ [16s/94s]
```

I have no idea what the time consuming parts are of the test suite, but if they are the HTTP requests you could explore mocking or caching - so funny 😆 i was just about to suggest trying `webmockr` with `curl` and remembered you opened this issue https://github.com/ropensci/webmockr/issues/40 - anyway, I'll try to get that webmockr/curl integration done soon.

- You can put a ropensci review badge in your README

- Goodpractice output. Nothing much to, nice work:

```r
── GP infx ────────────

It is good practice to

    R/dataset.R:515:NA
    R/feature.R:153:NA
    R/file.R:288:NA
    R/image.R:174:NA
    R/json-base.R:110:NA
    ... and 13 more lines
```

Seeking reviewers now 🕐

---

Reviewers: xx xx
Due date:
