### 2018-07-12 #218 @ottlngr
Estimated hours spent reviewing: 3,5

---

First of all I have to state that `infx` is implemented in an exemplary manner. Though I have absolutely no biological background and having a hard time to imagine what the data `infx` gives access to actually is, I could easily follow the idea and structure of the package and the procedures it provides.

With detailed documentation and several vignettes, the potential user is guided from the very first steps until mastering the package.

While I let the check marks in the reviewer template speak for themselves, I will elaborate on some (minor) issues I came across.

- The package contains no code of conduct or contribution guidelines as demanded by the [rOpenSci Packaging Guide](https://ropensci.github.io/dev_guide/building.html#readme)
- Also the [rOpenSci badge](https://ropensci.github.io/dev_guide/building.html#readme) and the [repostatus badge](http://www.repostatus.org/) are missing
- As `infx` imports `magick`, potential users are confronted with a system dependency. According to the rOpenSci Packaging Guide such dependencies should be indicated in `DESCRIPTION` and might be checked using a configure script. Examples can be found in the [Packaging Guide](https://ropensci.github.io/dev_guide/building.html#package-dependencies)
- The use of the `codemetar` package could be worthwhile for the package

`infx` primarily seems to address academic users. The academic flair of the package is further emphasized by the fact the development of the package was funded. Nevertheless I would have enjoyed one or two sentences explaining purpose of the package

- for people without a distinct biological background, and/or
- for people being no native English speaker an struggling with the many technical terms.

Surely, this is the opinion of someone fulfilling both these criteria -but maybe it is an valuable input for the author.

Running `devtools::spell_check()` returns some, even if few typos and different spellings of certain terms.

As stated earlier, I think `infx` accomplishes the promised tasks in a remarkable manner.

Only one thing, which I always try to eliminate when writing a package or functions dealing with APIs, attracted my attention:

```
token <- login_openbis()
projects <- list_projects(token)
adeno_exps <- list_experiments(token, projects[[1L]])
```

As a user I find it redundant to supply the `token` to `list_experiments()`, when already supplying `projects` which was received using the `token`. This might be a cosmetic issue, but maybe there is a way to store the token within the respective objects, such as `projects`.

Furthermore `infx` uses `curl` to handle HTTP requests. Maybe the `httr` package can be used to meet the recommendation of rOpenSci.

I congratulate @nbenn on a valuable contribution to the R community and hope, my review is felt as, even if from a non-specialist, valuable input.
