### 2018-07-16 #218 @nbenn
@ottlngr: thank you for your kind words.

I fixed the missing community guidelines and repo badges and added a codemeta description. I also fixed some spelling errors using `devtools::spell_check()`. I didn't know about this, thank you for the tip!

As for the ImageMagick dependency, I added it as `SystemRequirements` field in the `DESCRIPTION` file. But is this really what one should do? After all, it is not a dependency of `infx`, but of `magick`. The same goes for a `configure` script: I'm happy to copy the relevant parts from `magick` but this only leads to the same checks being run twice. @sckott, how do you feel about this? Is it customary to list and/or check for downstream system dependencies?

I tried to give a some more context to this package in the beginning of the readme. Does this help a bit for people coming from other fields?

I did consider using `httr`, but as far as I understand, `httr` doesn't support async requests (r-lib/httr#271). This is why I opted for `curl` instead.

Finally, I like your suggestion regarding token handling. I could easily attach the used token as `token` attribute to objects returned from the API. I'd need some additional logic to handle subsetting, concatenating etc. of objects. Also expired tokens need to be handled. I'll think about this.

Thank you for your feedback @ottlngr.
