### 2018-07-16 #218 @ottlngr
> As for the ImageMagick dependency, I added it as SystemRequirements field in the DESCRIPTION file. But is this really what one should do? After all, it is not a dependency of infx, but of magick. The same goes for a configure script: I'm happy to copy the relevant parts from magick but this only leads to the same checks being run twice.

Indeed. The imported `magick` actually does handle the system dependency very well.

> I tried to give a some more context to this package in the beginning of the readme. Does this help a bit for people coming from other fields?

It does, yes!

> I did consider using httr, but as far as I understand, httr doesn't support async requests (r-lib/httr#271). This is why I opted for curl instead.

That is totally fine as long there's a specific reason.

> Finally, I like your suggestion regarding token handling. I could easily attach the used token as token attribute to objects returned from the API. I'd need some additional logic to handle subsetting, concatenating etc. of objects. Also expired tokens need to be handled. I'll think about this.

I think that would be cool, but I also see the necessary effort. Maybe just keep it in mind for a future release.

I added the final check marks in the reviewer template!
