### 2018-07-17 ropensci/infx#6 @nbenn
Add token to returned objects
As per suggestion by @ottlngr, add token as attribute to returned `json_class` objects such that the token can be omitted in API calls that re-use previously fetched objects. See ropensci/onboarding#218.
