### 2018-07-20 #218 @tdjames1
>##### For packages co-submitting to JOSS
>
>- [ ] The package has an **obvious research application** according to [JOSS's definition](http://joss.theoj.org/about#submission_requirements)
>
>The package contains a `paper.md` matching [JOSS's requirements](http://joss.theoj.org/about#paper_structure) with:
>
>- [ ] **A short summary** describing the high-level functionality of the software
>- [ ] **Authors:**  A list of authors with their affiliations
>- [ ] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [ ] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

---

`infx` is an interface enabling access to openBIS structured experimental data collected as part of the InfectX project.  At the core of the package is an interface to the JSON based openBIS API.  Further functionality facilitates access to data structures used in the InfectX experimental data.  As noted in documentation, the use of the package is not restricted to the InfectX openBIS instance, but the functionality/API coverage is geared up to that instance.

I'm not familiar with how other instances of the openBIS differ from the InfectX instance.  I wondered if there may be a case to split this into two packages, one containing core functionality and another to provide the specific functionality required for the InfectX project data.  However, if the InfectX functionality would be useful in other openBIS instances then a single package that can be extended probably makes more sense.

Overall, the package is well designed and all functions seem to work robustly when tested using the examples provided in the README and vignettes (I'm not familiar with the data structures so I have not tested the functionality much beyond this).  Caveats, such as openBIS functionality not covered or inconsistencies in the openBIS API, are well documented in function documentation and openBIS API vignette.

The following comments are for the most part clarifications to documentation and notes on ROpenSci packaging guidelines.  I started my review prior to the commits made in response to @ottlngr's review so apologies if I've made a note of anything already fixed!

Tests ran successfully but `goodpractice::gp()` complains about coverage:
```
── GP infx ─────────────────────────────────────────────────────────────────────

    R/dataset.R:160:NA
    R/dataset.R:161:NA
    R/dataset.R:162:NA
    R/dataset.R:163:NA
    R/dataset.R:164:NA
```
The highlighted lines are within S3 methods e.g. `list_datastore_urls.character`.  I have not been able to find any advice about what to do in the case of testing S3 methods so I don't know if you need to duplicate tests for all overloaded methods. Perhaps @sckott can advise here.

The amendments to the README that @nbenn made in response to @ottlngr's review are a great improvement.

* OpenBIS or openBIS?
* typo: is available form
* links with text "here": suggest replacing with text that specifies where _here_ is e.g. "a browser-based view of the data is available from the [InfectX openBIS data browser]()"
* typhimurium - needs italics
* link "[1]" to footnote text not active
* "This R package provides access to the openBIS JSON-RPC API..." - append sentence to the effect that it's specifically applied to InfectX data?
* Github -> GitHub

### Vignettes

N.B. I reviewed the online vignettes (https://nbenn.github.io/infx/articles/infx-intro.html etc).

I found the vignettes extremely helpful in guiding me through the functionality of the package as it relates to the data structures accessed through openBIS.

#### Introduction to infx

* para 2 typo: accessed form here (should be from?)
* Organizational concepts in openBIS, para 4 typo: preformed
* Searching in openBIS, para 4 "For a small example" -> "In this example"?
* Searching in openBIS, para 5 punctuation + typo?: This MaterialGeneric generic object
* Retrieving openBIS data resources, para 1 typo: is considered data resources
* Retrieving openBIS data resources, para 1 typo: teated
* Image access, para 3 typo:  same sample of object

#### OpenBIS API coverage

* Creating a REST call, para 1 define REST?
* Creating a REST call, para 1 typo: the same input produce identical
* Creating a REST call, para 2 will be base::eval()ued -> will be evaluated using base::eval() ?
* Creating a JSON-RPC request, para 1 capitalisation of e.g. String, Structured value etc as per JSON-RPC specification is a bit confusing, is it necessary?
* Creating a JSON-RPC request, para 3 typo: if of length 1,
* Creating a JSON-RPC request: Should the examples for `make_request()` and `make_requests()` each create _and_ destroy login token?
* Creating a JSON-RPC request, para 8 capitalise API?: api endpoint url
* Creating a JSON-RPC request, para 8 typo: this makes possible to
* Creating a JSON-RPC request, para 9 typo: an url
* Creating a JSON-RPC request: In openBIS demo example, `logout_openbis(token)` gets separated from the rest of the demo code by the image output so when copying the code to clipboard it's easy to forget to run that line
* Summary of API methods formatting error?: ###Interface IDssServiceRpcGeneric etc

#### JSON object handling

* Creating `json_class` objects, para 4: to create json_class object (insert "a"?) + full stop at end of para
* In the constructed json_class example `identical(construct, json_class)` doesn't make much sense because the construct is a `foobar` example while the `json_class` object was set up to mimic openBIS data.
* In the `is_json_class` example it might be helpful to point out what is wrong in the test object.
* Creating `json_class` objects, para 7 typo: ASCI
* Using vectors of `json_class` objects, para 4 typo: and it alias

### Function documentation

  * Make cross references to other functions (e.g. in details for `list_samples`, "see `list_plates` and `list_wells`") into functioning links within R documentation.
https://cran.r-project.org/doc/manuals/r-release/R-exts.html#Cross_002dreferences

  * Make links to openBIS API documentation display correctly as links within R documentation.

  * Change heading "TODO" to e.g. "Implementation note"?

* ROpenSci package guidelines:

> _All functions should document the type of object returned under the @return heading._

> _We recommend using the @family tag in the documentation of functions to allow their grouping in the documentation of the installed package._

#### `infx`
* typo: form (instead of from, in a couple of places)
* deduplication?
* links to [openBIS API vignette], [section on JSON objects] and [JSON object vignette] don't work
* Modify introductory text as in README to provide a bit more context? ROpenSci package guidelines:

> _As is the case for a README, top-level documentation or vignettes may be the first point of entry for users. If your package connects to a data source or online service, or wraps other software, it should provide enough information for users to understand the nature of the data, service, or software, and provide links to other relevant data and documentation._

* Add link to "Introduction to infx" vignette?

#### `print.json_class`
* Punctuation in final para
* typo: with/length

#### `json_vec`
* typo: is is some cases

#### `list_datasets`
* typo: is a that of
* typo: fetched
* typo: may returned by

#### `list_files`

  * typo: to fetch_files()), All available
  * "Files can only be retrieved after previously..." - join first 2 sentences here?

#### `list_image_metadata`

  * typo: by passing a objects
  * typo: it it is set to format

#### `list_materials`

  * arguments > x: MaterialIdentifier should be in code typeface

#### `list_samples`

  * Missing closing bracket at end of first sentence.
  * typo: and for and wells
  * Last sentence is unclear.

#### `make_requests`

  * Due to the line wrapping the following section of the documentation gets formatted as the start of a list; rephrase or rearrange? I'm not completely clear what "is on length 1" means anyway.
```
...such that the topmost list level is on length
#' 1. The function
```

  * Missing space: to`json_class`
  * typo: if an objects is used
  * Should the following refer to `make_requests()` rather than `make_request()`?
```
# or using make_request(), the params argument has to be a list per
# request and the first entry of the returned list has to be selected
```
  * typo: httbin

#### `search`

  * typo: consist of a desired value
  * typo:  a set of object of the
  * typo:
```
# of using the perm_id attribute
```

#### `list_datastores`

  * typo: Data store sever url
  * Running the "let timeout expire" example works for me if I send the block of code to the console - increase time in Sys.sleep() call?

### CRAN gotchas from ROpenSci package guidelines
* _Make sure your package title is in Title Case._
* _Make sure you include links to websites if you wrap a web API, scrape data from a site, etc. in the Description field of your DESCRIPTION file._
* _Avoid long running tests and examples. Consider testthat::skip_on_cran in tests to skip things that take a long time but still test them locally and on Travis._
