### 2018-07-23 #218 @sckott
@nbenn reviews are both in now, continue discussion here. let me know if you have any questions about process
