### 2018-07-25 ropensci/infx#8 @nbenn
Fix links from function docs to vignettes
Currently links from function docs to vignettes only work locally, but not in the pkgdown page. Perhaps this will be fixed (r-lib/pkgdown#772), otherwise they might have to be removed altogether.
