### 2018-07-25 #218 @nbenn
@tdjames thank you very much for for the time you took for your in-depth feedback. I fixed most of the issues you found, except for the following, where I either disagree, or do not understand.

### Re Review Comments
I considered advertising my package more as an openBIS API package, rather than an InfectX data access package, but as the InfectX data set is (afaik) the only instance of large-scale, publicly accessible data under openBIS, I find the data access more attractive than the API capabilities per-se. In principle, nothing in the package is specific to InfectX (with exception of the default values in `login_openbis()` and examples used in docs). Personally, I feel the focus on InfectX data is ok, as long as no other openBIS instances of comparable scope start showing up. Or do you have strong feelings otherwise?

### Re Tests
Test coverage is 98%, when I run tests locally or on Travis (also see [sckott's gp output](#issuecomment-396415373)). I don't understand why your gp output reports only 56% coverage. However I don't think low test coverage is an issue here. If anything, I have too many tests that result in lengthy runtimes. Therefore the problem I'd rather look into, is reducing test runtime using @scott's mocking infrastructure. Unfortunately I haven't gotten to this yet.

### Re README
* typhimurium is romanized on purpose, as typhimurium does not specify a species but a [serovar](https://en.wikipedia.org/wiki/Salmonella_enterica#Nomenclature)
* In light of what I tried to elaborate a bit above, what do you think needs changing here?
  > "This R package provides access to the openBIS JSON-RPC API..." - append sentence to the effect that it's specifically applied to InfectX data?

  Implementation wise, nothing is specific to InfectX data, it is more that the package presentation is built around this dataset.
* > link "[1]" to footnote text not active

  I tried to fix this, but unfortunately I think it's not possible, as there is no proper support for footnotes in gfm

### Re Vignettes
* in *OpenBIS API coverage*, I don't understand

  > Creating a JSON-RPC request: Should the examples for make_request() and make_requests() each create and destroy login token?

  A login token is created for demonstrating `make_request()` and `make_requests()`. This login token is destroyed because the following example is a demonstration of how to access a non-InfecX openBIS instance. Are you suggesting to create individual login tokens for both the `make_request()` and `make_requests()` examples?

### Re Function documentation
* I'm afraid I don't follow:

  > Make cross references to other functions (e.g. in details for list_samples, "see list_plates and list_wells") into functioning links within R documentation.
https://cran.r-project.org/doc/manuals/r-release/R-exts.html#Cross_002dreferences

  For me, both on the pkgdown website and in the R Documentation pane of RStudio, `list_plates()`, etc. in `?list_samples` are clickable links. I think that under roxygen2 with markdown support enabled, the correct syntax is `[list_plates()]` which is rendered to `\code{\link[=list_plates]{list_plates()}}`. Am I missing something here?

* Re ROpenSci package guidelines:

  > All functions should document the type of object returned under the `@return` heading.

  @sckott Is this a strict requirement? The way documentation is currently organized, in part to the heavy use of S3 generic methods, I find it hard to use `@return` tags in a sensible manner. Multiple `@return` entries in a file are just pasted together without the context of which class-specific method they belong to. Instead I opted to describe the possible return types in the *Details* sections. Is this also acceptable?

* Re infx
  > links to [openBIS API vignette], [section on JSON objects] and [JSON object vignette] don't work

  See nbenn/infx#8

### Re CRAN gotchas from ROpenSci package guidelines

* > Avoid long running tests and examples. Consider testthat::skip_on_cran in tests to skip things that take a long time but still test them locally and on Travis.

  I think the way I set up my tests, the majority of the test suite will be skipped when
  - on travis but not the release build
  - on appveyor
  - on CRAN

  Do you have reason to believe my `check_skip()` function in `tests/testthat/setup-functions.R` will not work properly on cran?
