### 2018-07-25 #218 @sckott
- for `@return`: can you show me some examples?
- the cran check_skip fxn looks good
