### 2018-07-29 #218 @nbenn
@sckott, sorry for the delay. I started writing up a lengthy answer, explaining what I meant regrading the `@return` tags. In the process of doing so, I decided on simply adding the `@return` sections instead. I'll get back to you when this is done.
