### 2018-07-29 #218 @tdjames1
Hi @nbenn, apologies for the delay in responding to your queries/comments. Please find my responses below.

> Re Review Comments
>
> Personally, I feel the focus on InfectX data is ok, as long as no other openBIS instances of comparable scope start showing up. Or do you have strong feelings otherwise?

That's fine, I felt that I should raise it as an issue but I don't feel strongly that it is a big problem.

> Re Tests
>
> Test coverage is 98%, when I run tests locally or on Travis (also see sckott's gp output). I don't understand why your gp output reports only 56% coverage.

I think perhaps something is not set up correctly on my local machine. 98% seems pretty decent coverage!

> Re README
>
> typhimurium is romanized on purpose, as typhimurium does not specify a species but a serovar

OK! I didn't realise that.

>
> In light of what I tried to elaborate a bit above, what do you think needs changing here?
>
> "This R package provides access to the openBIS JSON-RPC API..." - append sentence to the effect that it's specifically applied to InfectX data?
>
> Implementation wise, nothing is specific to InfectX data, it is more that the package presentation is built around this dataset.

I think I got a different impression when I was reading some of the documentation, i.e. that in InfectX data is structured in a certain way which could be different in other instances. You could perhaps append something like "with specific reference to the InfectX dataset" to clarify the relationship between the package, openBIS and InfectX.

> link "[1]" to footnote text not active
>
> I tried to fix this, but unfortunately I think it's not possible, as there is no proper support for footnotes in gfm

That's a shame!

> Re Vignettes
>
> A login token is created for demonstrating make_request() and make_requests(). This login token is destroyed because the following example is a demonstration of how to access a non-InfecX openBIS instance. Are you suggesting to create individual login tokens for both the make_request() and make_requests() examples?

Yes, and I apologise for not being more explicit about this. In the vignette, each code block has a little "copy to clipboard" helper to make it easy for a reader to grab the code and paste it into R. But that also makes it easy to copy the first block which demonstrates `make_request()` but doesn't include the `logout_openbis()` call. That's why I thought it might be better if each block was complete in itself.

> Re Function documentation
>
> For me, both on the pkgdown website and in the R Documentation pane of RStudio, list_plates(), etc. in ?list_samples are clickable links. I think that under roxygen2 with markdown support enabled, the correct syntax is [list_plates()] which is rendered to \code{\link[=list_plates]{list_plates()}}. Am I missing something here?

It's displaying correctly for me now, so I'm not sure what happened there. Possibly another local configuration issue.

> Re CRAN gotchas from ROpenSci package guidelines
>
> Do you have reason to believe my check_skip() function in tests/testthat/setup-functions.R will not work properly on cran?

Apologies, I missed the `check_skip()` function. Looks good.
