### 2018-08-02 #218 @nbenn
Ok, the InfectX openBIS service is online again.

@sckott I added `@return` tags to the docs.

@tdjames1 I think, I now incorporated all the changes you suggested.
