### 2018-08-02 #218 @sckott
Approved!  Thanks again for your submission @nbenn !

* Please transfer the package  to `ropensci`- I've invited you to a team within our ropensci organization. After you accept you should be able to move it. You'll be made admin once you do.* Make sure to
* Make sure to
    * add rOpenSci footer to README
`[![ropensci_footer](https://ropensci.org/public_images/ropensci_footer.png)](https://ropensci.org)`
* In a new `.github` folder in the root of the pkg, add a `CONTRIBUTING.md` file, an issue template file, and a PR template file, see https://github.com/ropensci/dotgithubfiles/tree/master/dotgithub for examples, you can copy them directly and edit as you wish
* We're starting to roll out software metadata files to all ropensci pkgs via the Codemeta initiative, see  https://github.com/ropensci/codemetar/#codemetar  for how to include it in your pkg, after installing the pkg - should be easy as running `codemetar::write_codemeta()` in the root of your pkg
* you might want to add your ORCID id if you have one in the DESCRIPTION file, e.g. https://github.com/ropensci/taxize/blob/master/DESCRIPTION#L17

---

* Are you interested in doing a blog post for our blog https://ropensci.org/blog/ ?  either a short-form intro to it (https://ropensci.org/technotes/) or long-form post with more narrative about its development (https://ropensci.org/blog/). If so, we'll have our community manager @stefaniebutland get in touch with you on that
