### 2018-08-03 #218 @nbenn
@sckott @tdjames1 @ottlngr Thank you for taking the time to review and help me improve `infx`.
